+++ mediation/IronSource/source/plugin/Assets/GoogleMobileAds/Mediation/IronSource/Platforms/iOS/IronSource/IronSourceClient.cs
@@ -12,19 +12,16 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#if UNITY_ANDROID
+#if UNITY_IOS
 
 using UnityEngine;
-using System.Reflection;
 
-using GoogleMobileAds.Common.Mediation.IronSource;
+using GoogleMobileAds.Mediation.IronSource.Common;
 
-namespace GoogleMobileAds.Android.Mediation.IronSource
+namespace GoogleMobileAds.Mediation.IronSource.iOS
 {
     public class IronSourceClient : IIronSourceClient
     {
-        private const string IRONSOURCE_CLASS_NAME = "com.ironsource.mediationsdk.IronSource";
-
         private static IronSourceClient instance = new IronSourceClient();
         private IronSourceClient() { }
 
@@ -38,14 +35,12 @@ namespace GoogleMobileAds.Android.Mediation.IronSource
 
         public void SetConsent(bool consent)
         {
-            AndroidJavaClass ironSource = new AndroidJavaClass(IRONSOURCE_CLASS_NAME);
-            ironSource.CallStatic("setConsent", consent);
+            Externs.GADUMIronSourceSetConsent(consent);
         }
 
         public void SetMetaData(string key, string metaDataValue)
         {
-            AndroidJavaClass ironSource = new AndroidJavaClass(IRONSOURCE_CLASS_NAME);
-            ironSource.CallStatic("setMetaData", key, metaDataValue);
+            Externs.GADUMIronSourceSetMetaData(key, metaDataValue);
         }
     }
 }